+++ src/util.hpp
@@ -84,8 +84,8 @@ static inline T clamp(T min, T value, T max) {
     }
 }
 
-template <typename T, size_t n>
-constexpr size_t array_length(const T (&)[n]) {
+template <typename T, long n>
+constexpr long array_length(const T (&)[n]) {
     return n;
 }
